separate category and product getter
product getter now works for single products, ready to parse product objects
