+++ main.py
@@ -1,4 +1,4 @@
 # Script that pulls data from requests, processes it and saves it in a database
-import src.sodimac.sodimac_data_getter as sodimac_dg
+import src.sodimac.category_getter as sodimac_dg
 
 print(sodimac_dg.categories)
