+++ src/sodimac/category_getter.py
